third_party/WebKit/LayoutTests/fast/canvas-api/canvas-scroll-path-into-view-expected.txt

Series of tests to ensure correct results of scrolling path into view on canvas

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test case 1: scrollPathIntoView() / CTM == identity
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650


Test case 2: scrollPathIntoView() / CTM != identity
PASS testValue is 136
PASS testValue is 636
PASS testValue is 636
PASS testValue is 1136
PASS testValue is 636
PASS testValue is 127
PASS testValue is 627
PASS testValue is 627
PASS testValue is 1127
PASS testValue is 627
PASS testValue is 133
PASS testValue is 633
PASS testValue is 633
PASS testValue is 1133
PASS testValue is 633
PASS testValue is 160
PASS testValue is 660
PASS testValue is 660
PASS testValue is 1160
PASS testValue is 660


Test case 3: scrollPathIntoView(path2d) / CTM == identity
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650
PASS testValue is 150
PASS testValue is 650
PASS testValue is 650
PASS testValue is 1150
PASS testValue is 650


Test case 4: scrollPathIntoView(path2d) / CTM != identity
PASS testValue is 136
PASS testValue is 636
PASS testValue is 636
PASS testValue is 1136
PASS testValue is 636
PASS testValue is 127
PASS testValue is 627
PASS testValue is 627
PASS testValue is 1127
PASS testValue is 627
PASS testValue is 133
PASS testValue is 633
PASS testValue is 633
PASS testValue is 1133
PASS testValue is 633
PASS testValue is 160
PASS testValue is 660
PASS testValue is 660
PASS testValue is 1160
PASS testValue is 660


Test case 5: exceptions
PASS context.scrollPathIntoView(null); threw exception TypeError: Failed to execute 'scrollPathIntoView' on 'CanvasRenderingContext2D': parameter 1 is not of type 'Path2D'..
PASS context.scrollPathIntoView([]); threw exception TypeError: Failed to execute 'scrollPathIntoView' on 'CanvasRenderingContext2D': parameter 1 is not of type 'Path2D'..
PASS context.scrollPathIntoView({}); threw exception TypeError: Failed to execute 'scrollPathIntoView' on 'CanvasRenderingContext2D': parameter 1 is not of type 'Path2D'..


PASS successfullyParsed is true

TEST COMPLETE
